Refactor scripts to use service configs
